// src/app/Models/Trabajadores.ts
export interface Trabajadores {
    IDTrabajador?: number;
    NombreTrabajador: string;
    IDZona: string;
    IDArea: string;
}


export interface TrabajadoresTabla {
    IDTrabajador: number;
    NombreTrabajador: string;
    Zona: string;
    Area: string;
}


export interface TrabajadoresUsuario {
    idTrabajador: number;
    nombreTrabajador: string;
    zona: string;
    area: string;
  }
  